📦 Refactor common types
