+++ .changeset/orange-bears-leave.md
@@ -1,0 +1,9 @@
+---
+'@myst-theme/providers': patch
+'@myst-theme/common': patch
+'@myst-theme/article': patch
+'@myst-theme/site': patch
+'@myst-theme/book': patch
+---
+
+Move types to @myst-theme/common
