+++ packages/common/CHANGELOG.md
@@ -1,0 +1,3 @@
+# @myst-theme/common
+
+Common utilities and types for working with the MyST Theme and sites
