+++ packages/common/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@myst-theme/common",
+  "version": "0.4.0",
+  "type": "module",
+  "exports": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "license": "MIT",
+  "sideEffects": false,
+  "scripts": {
+    "clean": "rimraf dist",
+    "lint": "eslint \"src/**/*.ts*\" -c ./.eslintrc.cjs",
+    "lint:format": "prettier --check \"src/**/*.{ts,tsx,md}\"",
+    "test": "vitest run",
+    "test:watch": "vitest watch",
+    "build:esm": "tsc --project ./tsconfig.json --module es2015   --outDir dist --declaration",
+    "build": "npm-run-all -l clean -p build:esm"
+  },
+  "dependencies": {
+    "myst-common": "^1.1.0",
+    "myst-config": "^1.1.0",
+    "myst-spec-ext": "^1.1.0",
+    "nbtx": "^0.2.3",
+    "unist-util-select": "^4.0.3"
+  }
+}
