+++ packages/common/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig/esm.json",
+  "include": ["."],
+  "exclude": ["dist", "node_modules", "src/**/*.spec.ts", "tests"]
+}
