+++ packages/site/package.json
@@ -19,6 +19,7 @@
     "@myst-theme/diagrams": "^0.4.0",
     "@myst-theme/frontmatter": "^0.4.0",
     "@myst-theme/jupyter": "^0.4.0",
+    "@myst-theme/common": "^0.4.0",
     "@myst-theme/providers": "^0.4.0",
     "classnames": "^2.3.2",
     "lodash.throttle": "^4.1.1",
