+++ packages/site/src/index.ts
@@ -1,5 +1,4 @@
 export * from './utils';
-export * from './types';
 export * from './loaders';
 export * from './components';
 export * from './pages';
