+++ packages/site/src/loaders/index.ts
@@ -1,5 +1,3 @@
 export * from './errors.server';
-export * as cdn from './cdn.server';
-export * from './utils';
 export * from './links';
 export * from './theme.server';
