+++ themes/article/remix.config.dev.js
@@ -47,11 +47,7 @@ module.exports = {
     'credit-roles',
     'tex-to-typst',
     'jats-tags',
-    '@myst-theme/providers',
-    '@myst-theme/icons',
-    '@myst-theme/site',
-    '@myst-theme/jupyter',
-    '@myst-theme/frontmatter',
+    /^@myst-theme\/.*/,
     'react-syntax-highlighter',
   ],
   watchPaths: ['../../packages/**/*'],
