+++ themes/book/package.json
@@ -20,6 +20,7 @@
     "@myst-theme/jupyter": "^0.4.0",
     "@myst-theme/providers": "^0.4.0",
     "@myst-theme/site": "^0.4.0",
+    "@myst-theme/common": "^0.4.0",
     "@myst-theme/styles": "^0.4.0",
     "@remix-run/node": "~1.17.0",
     "@remix-run/react": "~1.17.0",
